requirements.txt
beautifulsoup4==4.12.0
bert-extractive-summarizer==0.10.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cssselect==1.2.0
cymem==2.0.7
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.7
huggingface-hub==0.13.3
idna==3.4
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
pydantic==1.10.7
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
scikit-learn==1.2.2
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
